backend/requirements.txt
Flask
Flask-PyMongo
Flask-Login
Flask-Mail
python-dotenv
dash
dash-bootstrap-components
numpy
pandas
plotly
PyPDF2
nltk
fpdf
openai
pypdf
docx
bson
certifi
ssl
azure-cognitiveservices-search-imagesearch